@@ -19,14 +19,19 @@ public class Conexion {
 
     private static final String JDBC_URL = "jdbc:mysql://localhost:3306/supercomprin?useSSL=false&useTimezone=true&serverTimezone=UTC&allowPublicKeyRetrieval=true";
     private static final String JDBC_USER = "root";
-    private static final String JDBC_PASSWORD = "123456789";
+    private static final String JDBC_PASSWORD = "";
 
     public static Connection getConnection() throws SQLException {
         return DriverManager.getConnection(JDBC_URL, JDBC_USER, JDBC_PASSWORD);
     }
 
     public static void close(ResultSet rs) throws SQLException {
-        rs.close();
+        try {
+            rs.close();
+        } catch (Exception e) {
+            System.out.println("Capturando error rs.close");
+        }
+        // rs.close();
     }
 
     public static void close(Statement smtm) throws SQLException {
